Move Symphony runner from sprites.dev to a Fly machine
sprites.dev's 30-second idle hibernation makes the D2 outbound-poll
Service we shipped in PR #383 effectively non-functional: the JS
event loop and any subprocess freeze the moment the machine
hibernates, and only an inbound network request will resume them.
We watched a single tick take 16+ minutes to pick up an issue
because nothing was knocking on the sprite from outside between
polls.

Switch to a Fly machine. The architecture is inverted: the runner
is now an HTTP server that processes one issue per `POST /tick`
request and exits the process after an idle window so the Fly
proxy stops the machine. GitHub Actions cron knocks every 15 min,
which auto-starts the machine. While takt is running, the long
HTTP request keeps the proxy from stopping us mid-run.

What's in the change:

- `infra/symphony/{Dockerfile,fly.toml,entrypoint.sh}`: Fly app
  config. Volume mounts at /data; entrypoint redirects $HOME to
  /data/home so claude / codex / cursor-agent / gh tokens persist
  across restarts and image rebuilds, the same way the sprites
  filesystem did.
- `bin/symphony-serve.ts`: HTTP server that owns the tick
  semantics. Bearer-token auth, single-issue concurrency, idle
  self-shutdown, SIGTERM forwarding to the takt child.
- `.github/workflows/symphony-tick.yml`: 15-min cron that POSTs
  /tick. Uses repo secrets `SYMPHONY_URL` + `SYMPHONY_TICK_TOKEN`.
- `docs/symphony-setup.md`: rewritten as the Fly setup runbook.
  The sprite-era runbook is preserved under
  `docs/symphony-setup-sprite.md` as a deprecated record of the
  earlier attempt — useful for the showcase write-up.
- `bin/symphony-runner.ts`: deleted. The sprite Service entry it
  served has no consumer now.
- `tests/structural/{no-format-in-batch,no-string-error,tilde-alias-import}.test.ts`:
  re-apply the 60s timeout that PR #377 introduced. Same flakiness
  on cold runs; the unmerged PR #377 carried the fix already, but
  we need it on this branch too.

Self-review pass per CLAUDE.md:

- Used the takt `## Judgment: COMPLETE` contract via
  `classifyTakt` (no new heuristics).
- SIGTERM is forwarded to the running takt child with a 10s grace
  before SIGKILL; the idle-shutdown timer is cleared on every tick.
- Empty / unparseable supervise output stays `failure_transient`
  rather than silently flipping to success.
- Bearer token is compared by exact string equality, only `/tick`
  and `/status` are routed.

Manual deployment steps (NOT automated, listed in the runbook):
1. `flyctl apps create upflow-symphony` + volume create
2. `flyctl secrets set SYMPHONY_TICK_TOKEN=...` + same value to
   `gh secret set SYMPHONY_TICK_TOKEN` and `SYMPHONY_URL`
3. `flyctl deploy` (first deploy lands the entrypoint waiting for
   auth)
4. `flyctl ssh console` then `gh auth login` /
   `claude auth login --claudeai` / `codex login` /
   `cursor-agent login` once
5. `flyctl machine restart` to drop into normal mode

Closes nothing yet — leaves the parent #370 open until end-to-end
runs through this stack.
