+++ infra/symphony/fly.toml
@@ -1,0 +1,47 @@
+# Symphony orchestrator host. Separate Fly app from upflow itself so it can
+# auto-stop independently and so a misbehaving runner can't take the main
+# app down. Deploy with:
+#   flyctl deploy -c infra/symphony/fly.toml --dockerfile infra/symphony/Dockerfile
+
+app = "upflow-symphony"
+primary_region = "nrt"
+kill_signal = "SIGTERM"
+kill_timeout = "30s"
+
+[build]
+  dockerfile = "Dockerfile"
+
+[env]
+  HOME = "/data/home"
+  SYMPHONY_REPO_DIR = "/data/upflow"
+  PORT = "8080"
+
+[[mounts]]
+  source = "symphony_data"
+  destination = "/data"
+
+[[services]]
+  protocol = "tcp"
+  internal_port = 8080
+  # The HTTP /tick handler runs takt synchronously so the Fly proxy sees an
+  # active connection for the entire 60-90 min run. We do NOT set
+  # auto_stop_machines because the proxy's "no traffic" heuristic fires on
+  # subprocess-only activity. The app self-exits after an idle window
+  # (see bin/symphony-serve.ts) and the proxy auto-starts on the next tick.
+  auto_stop_machines = "off"
+  auto_start_machines = true
+  min_machines_running = 0
+
+  [[services.ports]]
+    port = 80
+    handlers = ["http"]
+    force_https = true
+
+  [[services.ports]]
+    port = 443
+    handlers = ["tls", "http"]
+
+[[vm]]
+  cpu_kind = "shared"
+  cpus = 1
+  memory = "1gb"
