+++ tests/structural/no-format-in-batch.test.ts
@@ -109,7 +109,7 @@ describe('No `.format(...)` outside batch/helper/timeformat.ts', () => {
       allViolations.push(...findFormatCalls(abs))
     }
     expect(allViolations).toEqual([])
-  })
+  }, 60_000)
 
   it('catches a synthetic violation (sanity check on the matcher)', () => {
     const project = new Project({
